Differentiate our requests from WDTK requests
We need to be clearer about which requests were originated by us,
and which were originated on WDTK, since we can only send updates
to WDTK for requests we originated.

We add a new field request.remote_email, which contains the request
address for WDTK requests (and is null for others), and we no longer
set the remote_id for WDTK requests. Therefore the presence of the
remote_id may be used as an indicator that we can update the request
on WDTK.

TODO: send replies to the remote_email rather than the requestor
email, if there is a remote_email.