upgrade rails to 3.2.17
Have had to upgrade CarrierWave from 0.6.2 to 0.10.0 as well as the
older versions appear to no longer be available. Gems that would cause
a problem if newer versions were installed have been frozen.

All the tests are passing, but CarrierWave is causing a deprecation warning:

DEPRECATION WARNING: CarrierWave::MimeTypes is deprecated and will be removed
in the future, get the content_type from the SanitizedFile object directly.
(will be removed from version 0.11.0)