Raise an error if a new request from Alaveteli has a nil user_url
Requests which are loaded from the Alaveteli api are one of two kinds:
ones we know about already (because we sent them), and ones we don't
know about (because they were created on the Alaveteli instance).

When we're processing the second kind, it's possible that we know
about the user already (because they made a different request before),
or that they're new too. In either case, the user should have a
user_url, because they're a user of Alaveteli. The only time a user
should have an empty user_url is if they made the request on the
FOIReg site.

It should never be possible to get a user with no user_url that isn't
for a request we already know about, but because of a bug with the way
that requests are sent to Alaveteli, it was. When this happened, this
code ended up assigning the request to a different user, because it
would search for a user with external_url=nil, of which there are
many! ActiveRecord would helpfully return the first such user, causing
a confusing error.

Fixes #238