Patch Delayed::Worker to use ExceptionNotifier when jobs error
By default, when an exception is encountered in a delayed_job it just
re-queues the job and stores the error in it's database. This is ok,
but it means that we don't know when something goes wrong. Other
errors on the site send an email using ExceptionNotifier, so I've
made delayed_job do this too.

Closes #143