Increase the read_timeout for connections to Alaveteli
We had a situation, described in #237 where POST requests
to Alaveteli timed out, but as far as Alaveteli was
concerned, actually completed successfully. In the worst
case, they took nearly 90 seconds. This change raises the
timeout limit to 300 seconds in the hope that this is lone
enough for it to not happen again. Fixes #237