Add health checks
Almost all of this code was stolen from Alaveteli. My only additions
are the route for the health-check, the tests for the controller, and
the ExistenceCheck HealthCheck subclass.

This adds four health checks:
- new requests being created at least once every three days
- new requests being synced from Alaveteli at least once every three
  days
- there being no failed delayed jobs
- requests being sent to Alaveteli within an hour

These checks are run by hitting the /health url, which will return a
200 if everything is ok, and a 500 otherwise. The page prints a
summary of the checks in either case.

The checks which require connection to Alaveteli respect the
PUSH_TO_ALAVETELI and PULL_FROM_ALAVETELI config settings, but this
has required setting PULL_FROM_ALAVETELI to true in config/test.yml.
I don't think this has any effect on the rest of the code though.

Fixes #239