Set Delayed::Worker.max_attempts to 1
We have a known bug with delayed jobs that POST to alaveteli
instances and timeout but whose response is actually successful
(because Delayed::Job times out before the server responds).
We've increased the timeout, but it doesn't guarantee it'll never
happen. The main issue with timed out jobs is that they'll then
be re-tried, which results in duplicate requests on WDTK. As a
belt and braces fix, we reduce the max_attempts that Delayed::Job
will make to 1 here, so that we *never* get dupes under any
circumstances. This may cause lots of other failures that are very
annoying (because we'll have to manually reset the job to try again)
but we'll soon find that out!

This also updates a test of the retrying so that it tests that we
don't retry instead.